Worked on functions
Added a function for withdrawing transfercodes.
Edited the checkAvailability function to make sure a booking cannot be made with a later date as a check-in date.
Edited the sendBookingData function to send tourist, transfercode and total cost to bookings table.
Created a functionality for finding the ID of that booking.
Fixed the part of sendBookingData function which sends features into the booking_feature table,
Added the getCalendar function from calendar-test.php to functions.php